Install uncompressed kernel for debug/profiling purposes.
This installs vmlinux into /build/$BOARD/usr/lib/debug/boot. /usr/lib/debug is masked when building the
image, so the uncompressed kernel wont get installed on the final image. The STRIP_MASK prevents portage from
trying to run strip on it.

BUG=none
TEST=Built package. Checked that vmlinux is installed into the right place. Ran build_image and verified that
it is not installed onto the image. Made sure that portage doesn't try to strip the file.

Change-Id: I1c593de8f5c25772fec2c4798941de4ba47852a0
Reviewed-on: http://gerrit.chromium.org/gerrit/1808
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>