Install uncompressed kernel for debug/profiling purposes.
UPDATE: The first commit of this failed because I tested chromeos-kernel-next and assumed both ebuilds were
the same, which they werent =(. This time I tested both.

This installs vmlinux into /build/$BOARD/usr/lib/debug/boot. /usr/lib/debug is masked when building the
image, so the uncompressed kernel wont get installed on the final image. The STRIP_MASK prevents portage
from trying to run strip on it.

BUG=none
TEST=Built package. Checked that vmlinux is installed into the right place. Ran build_image and verified that
it is not installed onto the image. Made sure that portage doesn't try to strip the file.

Change-Id: Iebcc31dbc3fdc63b3302880b2fb0b8f95adb2428
Reviewed-on: http://gerrit.chromium.org/gerrit/1926
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>