bluez: disable consolekit USE flag in host environment
A previous CL disabled the consolekit USE flag for bluez in the target
environment as part of removing consolekit, but did not disable it in
the host environment. This causes missing dependencies when compiling
with USE=bluetooth because the bluetooth flag enables qemu-kvm's
bluetooth support which depends on the host version of bluez.

BUG=chromium-os:1809
TEST=USE=bluetooth ./build_packages

Change-Id: I7bb48d97f2f7a50b95f4b49aa18faed28b99b0cc
Reviewed-on: http://gerrit.chromium.org/gerrit/1985
Tested-by: Vince Laviano <vlaviano@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>