Nuke chroot hacks from before Gerrit transition.
These hacks are all at least two months old, and our official policy is to
remove hacks after two weeks. For future hacks, the plan is to use the chroot
upgrader script so that old chroots can be supported indefinitely.

Thanks to the Gerrit transition, there are probably few old chroots around
anyway that aren't going to be recreated anyway.

BUG=chromium-os:16151
TEST=Verify make_chroot, setup_board, and build_packages still work.

Change-Id: Ibde94a0e470b1b1fedd7da587600de0f719b2eb3
Reviewed-on: http://gerrit.chromium.org/gerrit/2156
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Tested-by: David James <davidjames@chromium.org>