Patch o3d to support pkgconfig sysroot prefixes again.
o3d has partial support for pkgconfig sysroot prefixes, but not all the
gyp files have the necessary magic to make it work. This patch updates
all the sysroot prefixes to work and updates the ebuild to make use of it.

This patch is needed so that we can upgrade pkgconfig to get rid of
hardcoded prefixes.

BUG=chromium-os:16242
TEST=Verify o3d compiles on arm and x86 both before and after pkgconfig
     change.

Change-Id: Icf3bc91d022c5dc2b57c71e6881baf76ec5e0ee7
Reviewed-on: http://gerrit.chromium.org/gerrit/2220
Reviewed-by: Anush Elangovan <anush@chromium.org>
Tested-by: David James <davidjames@chromium.org>
Reviewed-by: <zhurunz@google.com>