Make u-boot the default u-boot repository.
This CL does for u-boot ebuilds the same what
http://gerrit.chromium.org/gerrit/#change,1473 did for kernel
ebuilds. After this CL is introduced, the u-boot.git will be used
as the default repository, and using u-boot-next would require
the USE=u-boot-next setting.

chromeos-u-boot-build-env-9999.ebuild is an edited copy of
chromeos-u-boot-next-build-env-9999.ebuild (indicate use of
u-boot-next USE flag and make sure that the other package is
unmerged). The vboot_reference-firmware package is being modified
to depend on u-boot-next-build-env or u-boot-build-env, as
appropriate.

BUG=chromium-os:16224
TEST=manual
run
   emerge-tegra2_kaen u-boot
   USE=u-boot-next emerge-tegra2_kaen u-boot
   and observe different repositories used for building.

Change-Id: I3d2e6c23350ebb4b739996740d4b9a11404cb8a5
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/2249
Reviewed-by: David James <davidjames@chromium.org>