Get rid of most *.la files, and don't mangle *.pc files anymore.
This includes the following changes:
 - Update Portage to not mangle *.pc files.
 - Mask most libtool archive (.la) files so that they are not installed.
 - setup_board: generate a cross-pkgconfig script wrapper.
 - Remove ${ROOT} -I/-L from CFLAGS/LDFLAGS.
 - Update ImageMagick to set -I/-L flags because it gets confused by libtool.

The pkg-config wrapper will look into the board's root directory, and report all
paths relative to that. To work it requires a portage that no longer
plays with the paths in .pc files, which is not yet available.

Rather than using the standard pkg-config wrapper, though, use a
board-specific one. Setting the variable in make.conf should be enough
for ./configure and other build systems to pick it up.

BUG=chromium-os:9924, chromium-os:13441, chromium-os:14834, chromium-os:15595
TEST=Verify setup_board / build_packages work on x86, arm, and also internal
     targets

Change-Id: I932bd5f36beaf03258336aaac7c8c7978363573b
Reviewed-on: http://gerrit.chromium.org/gerrit/2180
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>