Allow toolchain developers to update pin before prebuilts are uploaded
Right now the toolchain developers are forced to update the prebuilts manually
before updating the toolchain pin and this has proved to be a bottleneck.
Instead we are going to trigger a toolchain build when the toolchain folks update
the pin. When the new prebuilts are ready, they will be uploaded automatically
and developers running setup_board with prebuilts will start using the new version.

I've also dropped the manual setup of PORTAGE_BINHOST, because the sdk builder now
uploads prebuilts.

BUG=chromium-os:16272
TEST=Test upgrade scenarios and verify that setup_board updates to latest prebuilts
     when --usepkg is used, and updates to pinned version when --nousepkg is used.

Change-Id: Ia8bfb23da9c0feea79368f6fea718f446298bbeb
Reviewed-on: http://gerrit.chromium.org/gerrit/2370
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Tested-by: David James <davidjames@chromium.org>