Fix circular dependencies in chroot builder.
This cleans up the chroot builder and moves to a new infrastructure similar
to our infrastructure used for boards.
  1. Packages are now built in our regular chroot using the --emptytree
     option.
  2. Once all the packages are built, they are installed to our board image,
     along with the tarballs used to build them.
  3. We then tar up the resulting image and use that to build a new chroot.

This new infrastructure ensures that all packages always have the dependencies
they need when they are built. Because we're building in the regular chroot,
the packages needed are already there so there are a lot less issues with
circular dependencies.

To get rid of the last circular dependency, I also needed to disable the
'git' use flag to gettext. This flag seems to be unimportant as gettext is
only used in our chroot and the 'git' use flag just controls whether git is
used to 'compress infrastructure archives.'

BUG=chromium-os:16421
TEST=Run chroot builder build.

Change-Id: If88dda924277aefb4272e081e2c2cf6479186236
Reviewed-on: http://gerrit.chromium.org/gerrit/2488
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>