Pin Chrome to chromeos-chrome-14.0.789.0_rc-r1 in the hopes to make tests stable.
This also moves the binhost back to the 789.  That's not pinned, but at least
people will get binary builds for a short period of time.

We still need to figure out what is causing the crash, but this should at least
keep the rest of Chrome OS able to develop until we do.

BUG=chromium-os:16451
TEST=Ran 'emerge-x86-generic -gp chromeos-chrome' and saw old chrome.

Change-Id: I0241a954906e4e5696e5ac051d6b5d63acca7e8e
Reviewed-on: http://gerrit.chromium.org/gerrit/2554
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>