Modified setup_board, adding --latest_toolchain option.
This adds a --latest_toolchain option to setup_board. This can be used by toolchain engineers
and buildbots to build with the latest experimental toolchain. Crossdev will choose the latest stable or
testing ebuild.

Also added glibc-2.11 to package.mask so that it is never used by crossdev.

BUG=chromium-os:16639
TEST=Ran setup_board with --latest_toolchain. Unmasking a newer toolchain ebuild causes it to be used when
running setup_board with --latest_toolchain.

Change-Id: Ifb77f8da68ad9f6e873ba8ab0bb2ce2b7e628b61
Reviewed-on: http://gerrit.chromium.org/gerrit/2686
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>