Fix candidate window & preedit to appear properly on the first click
This is a patch for intergrating the following two fixes.
http://codereview.appspot.com/4635044/
http://codereview.appspot.com/4635049/

BUG=chromium-os:16500
TEST=Open a page with a textd area, click it, and type a character in
CJK IME. Verify the candidate window immediately rendered below the
text are. For preedit area, apply codereview.chromium.org/7189008,
open a page with PPAPI plugin, click it, and type some character by
using IME. Verify preedit text is drawn on the candidate window.

Change-Id: Ib4f7569832108ace3c77637f6f9e283a7230a62e
Reviewed-on: http://gerrit.chromium.org/gerrit/2869
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Takayoshi Kochi <kochi@chromium.org>