Fix chromeos-chrome dependencies on python and flimflam-test
The chromeos-chrome package is missing dependencies on both python and
flimflam-test which are needed when USE=build_tests.

Python is required to build autotest.py (and possibly other parts as well).

flimflam-test is required because one of the tests built as part of the
chromeos-chrome build imports flimflam.  In order for Chromium to properly
build its test, the flimflam-test package needs to be installed so that
Chromium has access to it when building.

This has not been seen due to parallel emerge retrying failed builds
later after other packages were built on their own.

BUG=chromium-os:16731
TEST=In fresh chroot with chromeos-chrome-9999 unmasked:
  1) Run emerge-x86-generic -g chromeos-chrome
  2) Observe failure with no target for autotest.py
  3) Manually emerge python
  4) Observe failure due to missing flimflam module in chromeos-chrome
  5) Manually emerge flimflam-test
  6) Observe success in chromeos-chrome
  Confirming fix:
  7) Clear build environment (setup_board --force --board=x86-generic)
  8) emerge-x86-generic -g chromeos-chrome
  9) Observe success
  10)Run build_packages and build_image to confirm the full boot process works

Change-Id: Iec936a1c7cad7323514e790ba3402c69119596a8
Reviewed-on: http://gerrit.chromium.org/gerrit/2931
Tested-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>