u-boot-next: x86: temporarily remove dependency to vboot_reference-firmware
So far for the coreboot, it couldn't care less about verified boot. To
make the development easier, the dependency to verified boot is disabled
for now.

BUG=chromium-os:16813
TEST=manual

Run this to remove any dependencies and give you a clean environment for
testing:
  emerge-x86-alex -C \
    chromeos-u-boot-next-build-env \
    vboot_refernece-firmware \
    chromeos-u-boot-next

Run this and make sure that it builds and it does not depend on
vboot_reference:
  emerge-x86-alex chromeos-u-boot-next

Change-Id: If7db3bb176fcb3bc596e04b372b4d97d4d784d27
Reviewed-on: http://gerrit.chromium.org/gerrit/2956
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>