fdt: Make chromeos-bootimage pack device tree binary into image
The fdt blob is required for u-boot-next to run. This adds a new
variable used by the eclass, CROS_ARM_FIRMWARE_DTB, which specifies
the device tree to append to U-Boot. This is then set up by
the u-boot-next ebuild.

BUG=chromium-os:11623
TEST=emerge-tegra2_seaboard chromeos-bios; check it still boots ok into
stub, normal, recovery, developer

emerge-tegra2_seaboard chromeos-bootimage; check that it boots on Seaboard
(i.e. the device tree works).

Change-Id: I6c5a61cc21b0706e149a92423c9316695050cc68
Reviewed-on: http://gerrit.chromium.org/gerrit/3040
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>