Optimize chrontel driver for wm bogus_screen_resize use flag
This changes the ebuild to pass the bogus_screen_resize use flag
to the chrontel C code as -DBOGUS_SCREEN_RESIZE. This flag causes
the chromeos window manager to clear the unused areas of the screen
when responding to a (fake) root window resize event. The chrontel
code can use this behaviour to optimize screen resizing.

BUG=chrome-os-partner:4380
TEST=build using gmerge and see define on compiler command line

Change-Id: I47c0227f82709912a21c704a95a6763cc43a5172
Reviewed-on: http://gerrit.chromium.org/gerrit/3316
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Anush Elangovan <anush@chromium.org>
Tested-by: Mark Hayter <mdhayter@chromium.org>