Modify dtput to support strings with whitespace
This adds another patch to the series.

Previously dtput would use sscanf() to read the string from the command line,
even if no format string was provided. Thus it would split up words at
whitespace boundaries, and put a \0 between each.

This changes that behavior. Previously:

$ ../dtput dtget-test.dtb -ts /compatible this is a test

$ ../dtget dtget-test.dtb /compatible
116 104 105 115 0 105 115 0 97 110 111 116 104 101 114 0 116 101 115 116 0

and you have to use the -ts flag to see this as a list of strings

$ ../dtget dtget-test.dtb -ts /compatible
this is a test

Now:

$ ../dtput dtget-test.dtb -ts  /compatible "this is another test"

$ ../dtget dtget-test.dtb /compatible this is another test
this is another test

$ ../dtget dtget-test.dtb /compatible -tb
116 104 105 115 32 105 115 32 97 110 111 116 104 101 114 32 116 101 115 116 0

Note that the \0 terminators are gone.

BUG=chromium-os:17124
TEST=./run_tests with new test "Single string with spaces" added

Change-Id: I775e101a840626c5eb7d246cafc59d3f42a0ae97
Reviewed-on: http://gerrit.chromium.org/gerrit/3654
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Anton Staaf <robotboy@chromium.org>