Enable udev 170 on amd64.
This is a bit touchy as udev is used on both the 64 bit host and targets (and
udev 170 doesn't work on the host). Luckily we don't need udev at all on the
host, so we remove it from the host, so that this commit only adds it to 64
bit targets.

This brings us one step closer to converging x86 and amd64 targets.

TEST=ran make_chroot, setup_board, build_packages for x86-generic.

Change-Id: If25785e1645ff8deac69a348b8129f64817aa29f
Reviewed-on: http://gerrit.chromium.org/gerrit/4231
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>