Add an ebuild to build x86/coreboot/u-boot device trees.
In general, FDT ebuilds are broken into one ebuild for each logical group of
boards. For instance, the FDTs for all public tegra2 variants can be found here:

* http://git.chromium.org/gitweb/?p=chromiumos/overlays/board-overlays.git;
a=tree;f=overlay-tegra2/sys-boot/tegra2-public-firmware-fdts

This FDT build points at u-boot to find the source FDT. This is not a
requirement, but will probably continue to be how things work at least in the
short term. Since the FDT structure is under development, it is handy to have
the FDT and the code interpreting it in the same place.

The FDTs are not compiled / installed as part of the u-boot build to make it
easier / more obvious / more consistent for people that have private FDTs. In
this case it is expected that someone might put an FDT in a private board
overlay and not keep their FDT in the u-boot directory. Another reason that
FDTs aren't built as part of the u-boot build is that it keeps the u-boot build
from being board-specific, allowing us to share prebuilts among similar boards
in the future (even if those BOARDS have separate collections of FDTs)

On ARM, the corresponding ebuild is in a generic tegra2 overlay. On x86 there
isn't a similar generic overlay so it's being put here.

BUG=chrome-os-partner:4993
TEST=Emerged this ebuild.

Change-Id: I23fa0b4887a1af44dcf4eef1b9e5b9709076afa4
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4633
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>