xorg-server: Use static buffers to log from driver signal handlers
Back in the olden days, doing an xf86Msg() during a signal handler was a
big no-no, as it would deadlock the system on malloc when the prefix is
prepended.  Thus, only X_NONE messages could be used to bypass this.
However, LogVWrite limits to 1024 bytes anyway, so a patch was made to
eliminate malloc in LogVMessageVerb().

However, input driver messages these days go through xf86VIDrvMsgVerb(),
which does its own asprintf()/mallocs() to prefix the driver and device
names, reintroducing the same old no-no.

This patchset fixes this by introducing LogVHdrMessageVerb() and friends,
which allow a caller to specify a custom header as a variable string,
between the verb and the actually message.

In addition, by passing the header to the log layer, the verbosity checks
happen before any string processing, potentially speeding up processing
for production systems.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:18131
TEST=compiles for both 1.9.3 & 1.10
     xserver runs w/out crashing
     video driver messages in /var/log/Xorg.0.log with proper format
     input device driver messages in /var/log/Xorg.0.log with proper format

Change-Id: I1d5b51311f6e3321ceb4e2113fa28e435e48ea9a
Reviewed-on: http://gerrit.chromium.org/gerrit/4655
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>