kernel-headers: update version to be more meaningful
The kernel-headers version number is not meaningful, despite being
pinned to a particular version of the headers. Change to actually
state the appropriate kernel.

This means that there is no longer an excuse for ebuilds not declaring

  DEPEND=">=chromeos-base/kernel-headers-2.6.38"

for when they fail to build with earlier headers.

Likewise it means that the headers can be bumped (e.g. to 3.0) and a
package can declare a DEPEND to make sure it gets the upgrade.

BUG=none
TEST=verified emerge output

Change-Id: I1d3df610da1ed89de3153a03568f084b28b14b20
Reviewed-on: http://gerrit.chromium.org/gerrit/5354
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>