u-boot: obsolete 'grepping autoconf.mk' style of image generation
We were transitioning the build process from 'grepping autoconf.mk' to
cros_bundle_firmware. Since the transition is completed, we may now
retire the related scripts.

In addition, this patch makes cosmetic changes to variable names and
adds TODOs about the x86 image generation --- as we are not yet able to
generate images for x86 boards.

BUG=none
TEST=emerge-{x86-alex,tegra2_aebl} chromeos-bootimage

Change-Id: I684cb14f5c7e5db834c5acf10d841f2d25e5b522
Reviewed-on: http://gerrit.chromium.org/gerrit/5978
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>