resolve www35.vzw.com as localhost
resolve www35.vzw.com as localhost because it is never accessible
during modem activation when the modem is in a restricted IP pool.
www35.vzw.com is used for tracking cookies, and it takes Chrome a long
time to fail the synchronous requests to a host which never responds
to any IP packets.  Users should not be impact negatively by never
being able to access www35.vzw.com, and Verizon has been unable to add
www35.vzw.com to the restricted IP pool white list.

This should be reverted when we switch to using the new Verizon Portal
permanently see http://crosbug.com/21594

BUG=chrome-os-partner:4279,chromium-os:19421
TEST=try to activate with Verizon.  It should be faster to get to the part where they ask you for your zipcode.

Change-Id: I8704a086592eb05d958ecc16e20e794dd85bc26b
Reviewed-on: http://gerrit.chromium.org/gerrit/5580
Commit-Ready: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>