Added dev-install to base images
dev-install package is added to chromeos-base dependency list. It also
generates a list of packages that can be installed (which are in the dependency
list for chromeos-test and chromeos-dev but not chromeos). This list is used
later in the building process when uploading prebuilts.
Also add dhcp to the list of bootstrap packages. Emerging this package will not
work since group "dhcp" can not be created.

BUG=chromium-os:18178
TEST=Build base and test images successfully. Check contents of /etc/portage/
make.profile/package.installable.

Change-Id: I6d50eaa13979333e6df1d3c28f93cfbaff9a672b
Reviewed-on: http://gerrit.chromium.org/gerrit/5610
Reviewed-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>
Tested-by: Arkaitz Ruiz Alvarez <arkaitzr@chromium.org>