make_chroot: clean up timezone logic a bit more
Back in commit 6ac923fb6f1cfbff, the automatic timezone detection handling
was tweaked to account for possible transitional periods in a timezone
(i.e. daylight savings).  Unfortunately, one line was left behind which is
no longer used (probably due to the missed indentation).  So punt it.

Once that is gone, we can see that the first line is the same in both
branches of the if statement, so pull that duplication out.

Finally, fix incorrect indentation in the first branch (was using 3 spaces
instead of the standard 2).

BUG=None
TEST=deleted chroot, ran make_chroot, and verified chroot timezone w/`date`

Change-Id: I7faa233ee3610770f048b87282566887fdf0ce87
Reviewed-on: http://gerrit.chromium.org/gerrit/5653
Reviewed-by: David James <davidjames@chromium.org>
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>