autotest-tests: disable O3DSelenium tests
The test has been broken for several months, even before that, it has
been considered flaky. By stopping to build it in the default
distribution, I'd like to verify that nobody is using the test and will
be missing it. If that holds true, I'd like to remove this test
altogether, as it:
1) Doesn't work on arm: http://code.google.com/p/chromium-os/issues/detail?id=3270
2) Requires bulky java (icedtea), both for x86 boards and host chroot
3) In months, nobody stepped up to fix the test or even complained that it's broken.

BUG=none
TEST=none

Change-Id: I71b97d90a32ebffa56d2537a1032634003a3ae55
Reviewed-on: http://gerrit.chromium.org/gerrit/5687
Reviewed-by: Zdenek Behan <zbehan@chromium.org>
Tested-by: Zdenek Behan <zbehan@chromium.org>