make_chroot: drop LANG=C hardcoding
Some time ago, the LANG=C hardcoding was moved from make_chroot to
enter_chroot [1].  But it seems that there was some forking of code
bases (which were later merged) that caused make_chroot to gain this
hardcode again.  Now that we are back to one tree, and we don't run
apt in the chroot (it's Gentoo now, not deb based), and enter_chroot
takes care of hardcoding LANG=C, drop this setting from the default
user bashrc that gets setup by make_chroot.

[1] http://codereview.chromium.org/521053

BUG=None
TEST=deleted make_chroot, created a new one, saw LANG was still C upon enter

Change-Id: If84800ad525117e3061c7a20acfd0b0f9909c424
Reviewed-on: http://gerrit.chromium.org/gerrit/5761
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>