Let make continue building even if errors are encountered,
Sometimes there are several problems in the code causing compilation
errors, but the build system stops as soon as the first problem is
encountered, slowing down the process, as to see a problem in a
different source file one needs to start the build again.

Adding the `-k' make flag makes sure that make proceeds as far as
possible, compiling all source files but stopping before attempting to
link.

BUG=none
TEST=manual:
 emerged chromeos-kernel and chromeos-u-boot for alex and kaen
 platforms.

Change-Id: Ie107b73d11649ed6780e9acd07d5997b67e0331b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5789
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>