Turn off the mock TPM for x86-alex, stumpy, lumpy
We want to use the real TPM for x86-alex, stumpy and lumpy just like we are
for aebl and kaen. This change makes that happen by adding them to the white
list of boards where real TPM accesses are used.

BUG=chrome-os-partner:5431
TEST=Built vboot_reference-firmware with this ebuild, rebuilt coreboot/u-boot,
installed and booted them, ran vboot_twostop, and saw vboot attempt to access
the TPM where it hadn't before.

Change-Id: Ifd14a8fe9f077a87de1f5241ef5cd4a284c9d5b3
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/5878
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>