libcros/tpcontrol: Use tpcontrol_syncontrol if /dev/serio_raw* is present
Previously we were checking if the synaptics dbus pid file was present
to determine which tpcontrol_* script to use.
It is possible for synaptics userspace programs to be running (the
synaptics dbus client and even SynTPEnh), but the syntp driver is not
active since there is no /dev/serio_raw* device for it to open.
In these cases, don't use tpcontrol_syncontrol, instead try one of the
other methods to control a touchpad.

BUG=chromium-os:17285
TEST=(0) builds clean for x86
     (*) On system with syntp install as the active touchpad driver
     (1) Browse to chrome://system
     -> "touchpad" line should populate with all syntp properties
     (2) Browse to chrome://settings/system
     -> toggle "Enable tap-to-click", should work as expected
     (3) On a system with both syntp and cmt installed, disable syntp.
     -> There should be no /dev/serio_raw file
     (4) Browse to chrome://system
     -> "touchpad" line should populate with all cmt properties
     (5) Browse to chrome://settings/system
     -> toggle "Enable tap-to-click", should work as expected

Change-Id: I73b09dee1a0a91a38d9aacb05571ba3265d524a3
Reviewed-on: http://gerrit.chromium.org/gerrit/5997
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Dave Moore <davemoore@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>