Add system_api-9999.ebuild, which does nothing so far.
We'll be using the ebuild file to install files from system_api.git.

The previous attempt was reverted as it broke the build.
Before adding 9999.ebuild, system_api repository should have
been added to oldlayout.xml in both the internal and the
external manifest repositories.

This time, system_api repository has been added to oldlayout.xml
in the both repositories, so it'll be fine. system_api-0.0.1-r1.ebuild
will be created by the commit-bot.

BUG=chromium-os:19054
TEST=unmask x86 in KEYWORDS, and ran emrege-x86-alex system_api

Change-Id: I1b3974ac87437cc0b06d94ecb22e2650f162bb94
Reviewed-on: http://gerrit.chromium.org/gerrit/6017
Tested-by: Satoru Takabayashi <satorux@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>