Chrome build - make more directories writeable when using GERRIT_SOURCE
We need .repo to be writeable because because each project's .git
directory links to the .repo directory, so 'git checkout's will
fail otherwise.

Make more symlinks writeable so that we can accomodate changes to
the cros_DEPS/DEPS chrome DEPS file.

We will be removing all this addwrite code once we start
building off a copy of the source.

BUG=chromium-os:18386
TEST=Ran with CHROME_ORIGIN=GERRIT_SOURCE and SERVER_SOURCE

Change-Id: Ie1962c8a8fdd4285581576c4bcdd7d373d435d64
Reviewed-on: http://gerrit.chromium.org/gerrit/6455
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Ryan Cui <rcui@chromium.org>