chromium-os: relax the permissions error
Somehow our build system doesn't always update the permissions
on directories when it installs packages and so if you had the
r3 version emerged and try to emerge r4, the perms are set wrong
(which is a problem, but not likely to affect devs who do
incremental builds) and causes the entire build to fail.
Now just issue warnings which are easily overlooked because
warnings are the same color as 99.9% of all build messages.

Change-Id: I29aa5eb26c849e88f46bdc2869c31dfdf7379611

BUG=none
TEST=emerge with and without wrong perms

Change-Id: I25e8b4f7c3c743af62e737b6def3b24f24fdfc1e
Reviewed-on: http://gerrit.chromium.org/gerrit/6988
Reviewed-by: Ken Mixter <kmixter@chromium.org>
Tested-by: Ken Mixter <kmixter@chromium.org>