libscrypt: clean up ebuild
Since we're running a proper autoconf script, there is no need to run
any of the tc-getXX funcs.  So punt them.

The autoconf script, like any good little configure script, recognizes
CFLAGS/etc...  It does not recognize CCFLAGS at all.  So no point in
setting this -> punt it.

This library actually uses a func or two from openssl, so it needs it
at runtime, thus move the openssl dep into RDEPEND and DEPEND.

The ebuild currently doesn't install the static archive (just the
shared library), but it builds it anyways.  So add a USE=static-libs
to properly control that.

We've got IUSE=test, but no where do we actually `use test`, so punt!

For installing, we don't ever want to directly install the output of
libtool from its .libs/ subdir.  This often times can have bad rpath
tags in its ELF, or be in some other libtoolish intermediate step.
So use the proper `emake install` method so that libtool can do its
thing and we don't have to worry about it.

Finally, I tweaked the style a bit to be more "Gentooish".  Or at least
what I think is "Gentooish" :).

BUG=None
TEST=emerged it for x86/arm/amd64 and verified pkg contents before/after

Change-Id: Ia1d40854db9e5b31fbeeaa954558706e2b7fb8e9
Reviewed-on: http://gerrit.chromium.org/gerrit/7022
Reviewed-by: David James <davidjames@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>