laptop-mode-tools: wait for /usr to be mounted
The upstream version of laptop-mode-tools 1.59 depends on the file
hotplug.functions, which isn't present on all systems.  This patch
allows for systems which don't have it.

The dependency exists because etc/rules/lmt-udev calls wait_for_file in
hotplug.functions to wait for /usr to be mounted, before calling
/usr/sbin/laptop_mode.  On some systems, there is a delay and the
laptop_mode script may not be available right away.

This patch changes lmt-udev to first check for hotplug.functions.  If it
is not found, the script runs its own wait routine.  Thus, both systems
with and without hotplug.functions can run the script.

BUG=chromium-os:19865
TEST=Kill acpid using "stop acpid" so that AC plug/unplug events trigger
LMT through udev instead. Run "cat /sys/bus/usb/devices/*/power/control"
to see autosuspend state of usb devices.  Plug and unplug the power and
see if the autosuspend states change between on and auto.  If they do,
it means lmt-udev is running correctly.

Change-Id: Ic1f3d51c00cfaaf5501a9fb58ccb714581c3602a
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7026
Tested-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>