Fix the include and library paths passed to the compiler when building glog.
The is a bug in the glog Makefiles which causes the wrong header and library
paths to be passed to the compiler. This works around that problem by
substituting the correct values in directly. The long term solution is to
allow gtest-config and gmock-config work correctly with different sysroots.

BUG=chromium-os:19901
TEST=emerge-x86-alex glog && emerge-tegra2_kaen glog. Checked compile/link
lines. Also now works when host glibc is different from target libc.

Change-Id: I50148c95b144add8b91635ba26cc445862c79daa
Reviewed-on: http://gerrit.chromium.org/gerrit/7035
Reviewed-by: Vince Laviano <vlaviano@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>