Split the Google-supplied root certificates into a separate package.
This change creates a separate package, chromeos-base/root-certificates,
that provides just the trusted root certificates for Chrome OS.
The app-misc/ca-certificates package is restricted to installation
in the host chroot only and is not installed on the build target.

This change also reverts recent changes to the ca-certificates ebuild
that added error checks that are now no longer likely to be relevant.

BUG=None
TEST=Confirm /etc/ssl/certs was generated from chromeos-base/root-certificates
TEST=curl https://encrypted.google.com/
TEST=delete a board root, build without this change, apply this change, rebuild

Change-Id: Ica032ee57624c6112624e68948b1f4d72db9f42c
Reviewed-on: http://gerrit.chromium.org/gerrit/7080
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>