chromeos : remove chromeos-acpi dependency
Chromeos-acpi and acpid are not used for anything,
now that we are using udev for plug state changes.

BUG=chromium-os:19889
TEST=Build chromeos.
Ensure that /etc/acpi does not exist.
Make sure that plug/unplug events are still recognized
by powerd and by laptop mode tools for adjusting
backlight level and for enabling laptop mode
for usb device power management.

Change-Id: I5382877c9f2e58f3634a902c809193cab113a8c5
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7218
Reviewed-by: Simon Que <sque@chromium.org>