Modify mobile-providers ebuild to build and install libmobile-provider.
The mobile_provider API has been moved out of flimflam and into the
mobile-broadband-provider-info repo, so that it can be shared by
flimflam and shill. The unit tests and the browsedb tool have been
moved as well. Removed the test phase of the flimflam ebuild was
removed, since the mobile_provider unit tests were the only ones
that existed for flimflam.

BUG=chromium-os:19783
TEST=Built mobile-providers, with and without FEATURES=test, and
with and without USE=tools. Unit tests passed, and browsedb tool
was built and installed (when USE=tools). Also emerged and tested
flimflam.

Change-Id: I14fe160a1f771236127f1df1dae7783351615f77
Reviewed-on: http://gerrit.chromium.org/gerrit/7271
Tested-by: Eric Shienbrood <ers@chromium.org>
Reviewed-by: Eric Shienbrood <ers@chromium.org>