Allow chromeos-chrome ebuild to apply patches
This adds support to the chrome ebuild to apply local patches to any of the
code repositories used.  This is necessary as there are time when we need to
make fixes to P0 bugs while the patch is still being processed upstream.

No patches should be added to the ebuild that are not also in the process of
being upstreamed to the appropriate repository.

BUG=None
TEST=Built with patches empty, no change from current ebuild
  Added a patch protected by touchui_patches flag:
    Built once - saw patch applied and chrome built
    Built again - saw repo reset properly, reapply patch and build chrome
    Set -touchui_patches.  Saw patch reverted on build and not applied
  Built with -touchui_patches multiple times. Saw no increase in build time
    and patches were not applied and gclient reset not called

Change-Id: I5fbd46ea2375f406941c8b74de43ab60c9b97ae1
Reviewed-on: http://gerrit.chromium.org/gerrit/7330
Tested-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>