laptop-mode-tools: usb-autosuspend whitelist/blacklist in quotes
When the blacklist or whitelist is passed into the function
listed_by_id(), it should be in quotes, because the function expects to
get the entire list as one argument.  If it is not in quotes, each
element of the list will be passed in as a separate argument.

e.g. if you were passing in the list "0000:0000 1111:1111 2222:2222":
With quotes: $2 = "0000:0000 1111:1111 2222:2222"
Without quotes: $2 = "0000:0000", $3 = "1111:1111", $4 = "2222:2222"

BUG=chrome-os-partner:4041
TEST=Manually edit /etc/laptop-mode/conf.d/usb-autosuspend.conf.  Set
AUTOSUSPEND_USE_WHITELIST=1 and add more than one vendor:product ID pair
from the target system to AUTOSUSPEND_USBID_WHITELIST.  You can get the
vendor and product IDs by typing:
  cat /sys/bus/usb/devices/*/idVendor
  cat /sys/bus/usb/devices/*/idProduct
Now watch the autosuspend values of the usb devices:
  watch -n1 cat /sys/bus/usb/devices/*/power/control
Start on AC and unplug AC.  The devices that were included on the
whitelist should change from "on" to "auto".  The others should not.

Change-Id: If8675b05b6995a5b34dd99c777680e37b28cd5c3
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7340
Reviewed-by: Benson Leung <bleung@chromium.org>