Unmask glibc-2.11.1-r1 for host.
This updates glibc for the host. The githash has been updated to the latest
commit in the repository. PROVIDE= has also been set.

BUG=chromium-os:19436
TEST=I've done many builds with this change and the resulting chroot has been
good for building full images and booting them. In particular:
cros_sdk --bootstrap followed by building a full x86-alex and tegra2_kaen
image and booting them. Also, upgrading to this libc in an older chroot seems
to work correctly.

Change-Id: Id10df2e75deabe6f69e5d95aed2dc648cc54ff25
Reviewed-on: http://gerrit.chromium.org/gerrit/7581
Reviewed-by: asharif <asharif@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>