chromiumos-overlays: neuter pppd with nosystemconfig
When nosystemconfig is passed to ppp, upon negotiating IP addresses
in IPCP, it skips almost all of the final configuration, including
setting IP addresses for the ppp0, up'ing the interface, creating
routes (including default route), and writing resolv.conf.  This flag
is a superset of fetchpeerdns that jglasgow created, so remove that
older flag as it's not currently used.

Note one difference with nosystemconfig vs before is that we now
do sifnpmode(f->unit, PPP_IP, NPMODE_PASS) (which does PPPIOCSNPMODE)
before doing any kind of configuration of the interface.  And then
once flimflam is notified, flimflam does config in this order:
1) ups the interface
2) sets up local and peer addresses
3) sets up default gateway

whereas without nosystemconfig, NPMODE_PASS is sent to the driver
pppd does this config:
1) sets up local and peer addresses
2) ups the interface
3) sets up local and peer addresses (again)

The difference hasn't caused troubles in testing and we're leaving
the NPMODE_PASS stuff in pppd's world because we don't want flimflam
to know how to talk to ppp kernel modules.

TEST=network_VPN(OpenVPN&L2TP tests) + manual tests to RRAS server
BUG=chromium-os:17185

Change-Id: If8a226f23c089750428aa44b2f1a455f62c9bf48
Reviewed-on: http://gerrit.chromium.org/gerrit/7751
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Ken Mixter <kmixter@chromium.org>
Tested-by: Ken Mixter <kmixter@chromium.org>